+++ CHANGELOG.md
@@ -4,3 +4,14 @@
 
  - Added `File LambdaSystem::getUserDirectory()`.
  - Removed `ReturnAs`.
+
+## 1.3.0
+
+ - Added configuration API
+ - Added `Config`
+ - Added `FileConfig`
+ - Added `JsonConfig`
+ - Added `VirtualJsonConfig`
+ - Added `LambdaConstants.GSON`
+ - Added `LambdaConstants.GSON_PRETTY`
+ - Added `LambdaConstants.JSON_PARSER`